refactor: move computeStateForNextDay to its own file
