+++ src/game-logic/reducers/computeStateForNextDay.test.js
@@ -1,0 +1,56 @@
+import { shapeOf, testCrop } from '../../test-utils'
+import { generateCow } from '../../utils'
+
+import { computeStateForNextDay } from './computeStateForNextDay'
+
+jest.mock('../../data/maps')
+
+describe('computeStateForNextDay', () => {
+  beforeEach(() => {
+    jest.spyOn(Math, 'random').mockReturnValue(0.75)
+  })
+
+  test('computes state for next day', () => {
+    const {
+      cowForSale,
+      dayCount,
+      field: [firstRow],
+      valueAdjustments,
+      todaysNotifications,
+    } = computeStateForNextDay({
+      cowBreedingPen: { cowId1: null, cowId2: null, daysUntilBirth: -1 },
+      dayCount: 1,
+      field: [
+        [
+          testCrop({
+            itemId: 'sample-crop-1',
+            wasWateredToday: true,
+          }),
+        ],
+      ],
+      cowInventory: [],
+      historicalDailyLosses: [],
+      historicalDailyRevenue: [],
+      inventory: [],
+      itemsSold: {},
+      loanBalance: 0,
+      newDayNotifications: [],
+      notificationLog: [],
+      priceCrashes: {},
+      priceSurges: {},
+      profitabilityStreak: 0,
+      record7dayProfitAverage: 0,
+      recordProfitabilityStreak: 0,
+      todaysNotifications: [{ message: 'some message', severity: 'info' }],
+    })
+
+    expect(shapeOf(cowForSale)).toEqual(shapeOf(generateCow()))
+    expect(dayCount).toEqual(2)
+    expect(valueAdjustments['sample-crop-1']).toEqual(1.25)
+    expect(valueAdjustments['sample-crop-2']).toEqual(1.25)
+    expect(firstRow[0].wasWateredToday).toBe(false)
+    expect(firstRow[0].daysWatered).toBe(1)
+    expect(firstRow[0].daysOld).toBe(1)
+    expect(todaysNotifications).toBeEmpty()
+  })
+})
