+++ src/game-logic/reducers/index.test.js
@@ -1,4 +1,4 @@
-import { shapeOf, testCrop, testItem } from '../../test-utils'
+import { testCrop, testItem } from '../../test-utils'
 import {
   ACHIEVEMENT_COMPLETED,
   LOAN_INCREASED,
@@ -41,56 +41,6 @@ jest.mock('../../constants', () => ({
   CROW_CHANCE: 0,
   PRECIPITATION_CHANCE: 0,
 }))
-
-describe('computeStateForNextDay', () => {
-  beforeEach(() => {
-    jest.spyOn(Math, 'random').mockReturnValue(0.75)
-  })
-
-  test('computes state for next day', () => {
-    const {
-      cowForSale,
-      dayCount,
-      field: [firstRow],
-      valueAdjustments,
-      todaysNotifications,
-    } = fn.computeStateForNextDay({
-      cowBreedingPen: { cowId1: null, cowId2: null, daysUntilBirth: -1 },
-      dayCount: 1,
-      field: [
-        [
-          testCrop({
-            itemId: 'sample-crop-1',
-            wasWateredToday: true,
-          }),
-        ],
-      ],
-      cowInventory: [],
-      historicalDailyLosses: [],
-      historicalDailyRevenue: [],
-      inventory: [],
-      itemsSold: {},
-      loanBalance: 0,
-      newDayNotifications: [],
-      notificationLog: [],
-      priceCrashes: {},
-      priceSurges: {},
-      profitabilityStreak: 0,
-      record7dayProfitAverage: 0,
-      recordProfitabilityStreak: 0,
-      todaysNotifications: [{ message: 'some message', severity: 'info' }],
-    })
-
-    expect(shapeOf(cowForSale)).toEqual(shapeOf(generateCow()))
-    expect(dayCount).toEqual(2)
-    expect(valueAdjustments['sample-crop-1']).toEqual(1.25)
-    expect(valueAdjustments['sample-crop-2']).toEqual(1.25)
-    expect(firstRow[0].wasWateredToday).toBe(false)
-    expect(firstRow[0].daysWatered).toBe(1)
-    expect(firstRow[0].daysOld).toBe(1)
-    expect(todaysNotifications).toBeEmpty()
-  })
-})
 
 describe('purchaseItem', () => {
   describe('howMany === 0', () => {
